refactor: move icons [skip ci]
